fix: cast sample_rate to f32 to match seg.start/seg.end types
sherpa-onnx 1.13.4 OfflineSpeakerDiarizationSegment uses f32 for
start/end fields, not f64. This caused type mismatch errors when
computing sample offsets.

+++ apps/zhnote/src-tauri/src/stt.rs
@@ -166,8 +166,8 @@ impl SttEngine {
             let mut full_text = String::new();
 
             for seg in &diar_segments {
-                let start_sample = (seg.start * sample_rate as f64) as usize;
-                let end_sample = (seg.end * sample_rate as f64) as usize;
+                let start_sample = (seg.start * sample_rate as f32) as usize;
+                let end_sample = (seg.end * sample_rate as f32) as usize;
                 let end_sample = end_sample.min(samples.len());
 
                 if start_sample >= end_sample {
